Replace current system of remote & local Δs, partitions, revs,
temporary hacks for protocol response, with a proper event-based
design where project state itself is not recorded, events are and
are used for communication and append-only persistence.

+++ Code/webapp-base-test/shared/src/main/scala/shipreq/webapp/base/test/package.scala
@@ -11,7 +11,7 @@ package object test {
 
   trait TestDataId[D] {
     type I
-    def mkId(l: Long): I
+    def mkId(l: Int): I
     def setId(d: D, id: I): D
   }
 
@@ -22,7 +22,7 @@ package object test {
   type TestDataIdAux[D, Id] = TestDataId[D] {type I = Id}
 
   implicit object TagId_T extends TestObjDataId[Tag.type, Tag, TagId] {
-    override def mkId(l: Long) = ApplicableTagId(l)
+    override def mkId(l: Int) = ApplicableTagId(l)
     override def setId(t: Tag, i: TagId) = t match {
         case x: TagGroup      => x.copy(id = TagGroupId     (i.value))
         case x: ApplicableTag => x.copy(id = ApplicableTagId(i.value))
@@ -30,20 +30,20 @@ package object test {
   }
 
   implicit object ReqId_T extends TestObjDataId[ReqT.type, Req, ReqId] {
-    override def mkId(l: Long) = GenericReqId(l)
+    override def mkId(l: Int) = GenericReqId(l)
     override def setId(cf: Req, i: ReqId) = cf match {
         case r: GenericReq => r.copy(id = GenericReqId(i.value))
       }
   }
 
   implicit object CustomReqTypeId_T extends TestObjDataId[CustomReqType.type, CustomReqType, CustomReqTypeId] {
-    override def mkId(l: Long) = CustomReqTypeId(l)
+    override def mkId(l: Int) = CustomReqTypeId(l)
     override def setId(a: CustomReqType, b: CustomReqTypeId) = a.copy(id = b)
   }
 
   implicit object CustomFieldId_T extends TestObjDataId[CustomField.type, CustomField, CustomFieldId] {
     import CustomField._
-    override def mkId(l: Long) = Text.Id(l)
+    override def mkId(l: Int) = Text.Id(l)
     override def setId(cf: CustomField, i: CustomFieldId) = cf match {
         case f: Text        => f.copy(id = Text       .Id(i.value))
         case f: Tag         => f.copy(id = Tag        .Id(i.value))
@@ -52,7 +52,7 @@ package object test {
   }
 
   implicit object CustomIssueTypeId_T extends TestObjDataId[CustomIssueType.type, CustomIssueType, CustomIssueTypeId] {
-    override def mkId(l: Long) = CustomIssueTypeId(l)
+    override def mkId(l: Int) = CustomIssueTypeId(l)
     override def setId(a: CustomIssueType, b: CustomIssueTypeId) = a.copy(id = b)
   }
 }
